frontend: moved App to components folder
