+++ frontend/src/components/App.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import App from "./App";
+import App from "../../src/components/App";
 
 it("renders without crashing", () => {
   render(<App />);
